+++ src/index.ts
@@ -88,6 +88,17 @@ export {
   TabPanel,
 } from "./components/Tabs/Tabs";
 export {
+  TemplateAppContainer,
+  Template,
+  TemplateHeader,
+  TemplateBreakout,
+  TemplateContent,
+  TemplateContentTop,
+  TemplateContentPrimary,
+  TemplateContentSidebar,
+  TemplateFooter,
+} from "./components/Template/Template";
+export {
   default as TextInput,
   TextInputRefType,
 } from "./components/TextInput/TextInput";